Release v0.1.0 (#10)
* updating the charts to include the default images for infracost anc checkov
* updating the deployment to include the options on the controller flags
* updating the images as we no longer use the executor image to include all things
* adding a logging and recovery middleware to the apiserver
* updating the various utility packages
* updating the configuration controller to pass the images and options
* adding the images, command line options and the terraform version on the configuration
* adding the step command
* switching over to the job template
* fixing up the github workflow and the docker image
* fixing the linting
* increasing the timeout for watching for the job to start
* checking the KUBE_NAMESPACE has been set
* increasing the backoff on the watcher job
* adding the e2e workflow and linking it to the release
* adding some bats e2e checks to ensure the release

* [FEATURE] - Custom Job Override (#11)

* [FEATURE] - Custom Job Override

Adding the ability to override the job template

* Bump Version (#14)

Bumping the versions in the chart ready to release

* Allow for controller-kind errors (#15)

Update the Makefile to allow for errors in the controller-kind, as the stage is used to load the
images into the kind cluster for testing before the helm chart has been deployed

* - dropping the docs for the executor as it's no longer required

+++ pkg/controller/configuration/delete.go
@@ -73,8 +73,14 @@ func (c *Controller) ensureTerraformDestroy(configuration *terraformv1alphav1.Co
 		// @step: generate the destroy job
 		batch := jobs.New(configuration, state.provider)
 		runner, err := batch.NewTerraformDestroy(jobs.Options{
-			ExecutorImage: c.ExecutorImage,
-			Namespace:     c.JobNamespace,
+			EnableInfraCosts: c.EnableInfracosts,
+			ExecutorImage:    c.ExecutorImage,
+			InfracostsImage:  c.InfracostsImage,
+			InfracostsSecret: c.InfracostsSecretName,
+			Namespace:        c.JobNamespace,
+			ServiceAccount:   "terraform-controller",
+			Template:         state.jobTemplate,
+			TerraformImage:   GetTerraformImage(configuration, c.TerraformImage),
 		})
 		if err != nil {
 			cond.Failed(err, "Failed to create the terraform destroy job")
@@ -84,16 +90,18 @@ func (c *Controller) ensureTerraformDestroy(configuration *terraformv1alphav1.Co
 
 		// @step: we can requeue or move on depending on the status
 		if !found {
-			if err := c.CreateWatcher(ctx, configuration, terraformv1alphav1.StageTerraformDestroy); err != nil {
-				cond.Failed(err, "Failed to create the terraform destroy watcher")
+			if c.EnableWatchers {
+				if err := c.CreateWatcher(ctx, configuration, terraformv1alphav1.StageTerraformDestroy); err != nil {
+					cond.Failed(err, "Failed to create the terraform destroy watcher")
 
-				return reconcile.Result{}, err
-			}
+					return reconcile.Result{}, err
+				}
 
-			if err := c.cc.Create(ctx, runner); err != nil {
-				cond.Failed(err, "Failed to create the terraform destroy job")
+				if err := c.cc.Create(ctx, runner); err != nil {
+					cond.Failed(err, "Failed to create the terraform destroy job")
 
-				return reconcile.Result{}, err
+					return reconcile.Result{}, err
+				}
 			}
 
 			return reconcile.Result{RequeueAfter: 5 * time.Second}, nil
